Move some shared types out into a new tunnels_lib crate.
