+++ tunnels/Cargo.toml
@@ -15,3 +15,5 @@ midir = "^0.7"
 simple-error = "^0.2"
 lazy_static = "^1"
 typed_index_derive = "0.1.4"
+zmq = "0.9"
+tunnels_lib = { path = "../tunnels_lib" }
