+++ tunnels/src/look.rs
@@ -1,6 +1,7 @@
-use crate::{clock::ClockBank, mixer::Channel, numbers::UnipolarFloat, tunnel::ArcSegment};
+use crate::{clock::ClockBank, mixer::Channel, numbers::UnipolarFloat};
 use serde::{Deserialize, Serialize};
 use std::time::Duration;
+use tunnels_lib::ArcSegment;
 
 /// A look is a beam that is essentially the contents of an entire mixer.
 /// All channel settings are preserved.
