+++ tunnels_lib/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tunnels_lib"
+version = "0.1.0"
+authors = ["general electrix <general.electrix@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
